+++ packages/core-js/modules/esnext.typed-array.with-at.js
@@ -1,24 +1,13 @@
 'use strict';
-var lengthOfArrayLike = require('../internals/length-of-array-like');
-var isIntegralNumber = require('../internals/is-integral-number');
+var arrayWithAt = require('../internals/array-with-at');
 var ArrayBufferViewCore = require('../internals/array-buffer-view-core');
 
 var aTypedArray = ArrayBufferViewCore.aTypedArray;
 var exportTypedArrayMethod = ArrayBufferViewCore.exportTypedArrayMethod;
 var TYPED_ARRAY_CONSTRUCTOR = ArrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
 
-var ERROR_MESSAGE = 'Incorrect index';
-
 // `%TypedArray%.prototype.withAt` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-array.prototype.withAt
+// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.withAt
 exportTypedArrayMethod('withAt', function withAt(index, value) {
-  var O = aTypedArray(this);
-  var len = lengthOfArrayLike(O);
-  if (!isIntegralNumber(index) || index >= len) throw RangeError(ERROR_MESSAGE);
-  var actualIndex = index < 0 ? len + index : index;
-  if (actualIndex < 0) throw RangeError(ERROR_MESSAGE);
-  var A = new O[TYPED_ARRAY_CONSTRUCTOR](len);
-  var k = 0;
-  for (; k < len; k++) A[k] = k === actualIndex ? value : O[k];
-  return A;
+  return arrayWithAt(aTypedArray(this), this[TYPED_ARRAY_CONSTRUCTOR], index, value);
 });
