+++ packages/core-js/modules/esnext.typed-array.with-reversed.js
@@ -1,5 +1,5 @@
 'use strict';
-var lengthOfArrayLike = require('../internals/length-of-array-like');
+var arrayWithReversed = require('../internals/array-with-reversed');
 var ArrayBufferViewCore = require('../internals/array-buffer-view-core');
 
 var aTypedArray = ArrayBufferViewCore.aTypedArray;
@@ -7,12 +7,7 @@ var exportTypedArrayMethod = ArrayBufferViewCore.exportTypedArrayMethod;
 var TYPED_ARRAY_CONSTRUCTOR = ArrayBufferViewCore.TYPED_ARRAY_CONSTRUCTOR;
 
 // `%TypedArray%.prototype.withReversed` method
-// https://tc39.es/proposal-change-array-by-copy/#sec-array.prototype.withReversed
+// https://tc39.es/proposal-change-array-by-copy/#sec-%typedarray%.prototype.withReversed
 exportTypedArrayMethod('withReversed', function withReversed() {
-  var O = aTypedArray(this);
-  var len = lengthOfArrayLike(O);
-  var A = new O[TYPED_ARRAY_CONSTRUCTOR](len);
-  var k = 0;
-  for (; k < len; k++) A[k] = O[len - k - 1];
-  return A;
+  return arrayWithReversed(aTypedArray(this), this[TYPED_ARRAY_CONSTRUCTOR]);
 });
